// src/api/stetsom/model/productImage.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */

export interface ProductImage {
  image_id: string;
  library_id: string;
  /** @nullable */
  image_url?: string | null;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  order: number;
  created_at: string;
}